refactor: split stream_orchestrator tests into focused submodules [OPE-490]
Split the 606-line monolithic tests.rs into three focused modules:
- mocks.rs: all mock responder implementations (MockResponder,
  FailingUpdateResponder, FailingCreateResponder, FailingFinalizeResponder)
- success.rs: happy-path streaming tests (basic, throttled updates,
  truncation, discord intermediate updates, concurrent streams,
  many small deltas, telegram throttle, unicode content)
- failure.rs: failure and edge-case tests (sender dropped, error chunk,
  empty sender dropped, error empty buffer, lagged receiver, update
  failure, create draft failure, finalize failure, error truncation)

All 18 existing tests pass. No behavior changes.
