+++ crates/opengoose-core/src/stream_orchestrator/tests/mod.rs
@@ -1,0 +1,3 @@
+mod failure;
+mod mocks;
+mod success;
